Update z1top.v
This PR aims to make both the adder and counter circuit synthesizable at once, reducing overall time taken for a checkoff. (Currently, students need to rerun make program as the toggle between the adder and counter switch is done via a `#define` and an `#ifdef` in z1top.v)

The second switch (`SWITCHES[1]`) is used to toggle the output between the counter and adder circuit. This means that on the adder circuit, only the 4 buttons(`BUTTONS[3:0]`) can be used, reducing the maximum number of output bits(2'b + 2'b instead of 3'b + 3'b), with a maximum output of 3 bits instead of 4 bits. However, I do think this tradeoff is worth it due to the time cut from checkoffs.